// src/client/render/types/Renderer.ts
export enum TrainType {
  Engine = 0,
  TailEngine = 1,
  Carriage = 2,
}
// ...
export interface UnitState {
  id: number;
  unitType: string;
  ownerID: number;
  lastOwnerID: number | null;
  pos: number;
  lastPos: number;
  isActive: boolean;
  reachedTarget: boolean;
  retreating: boolean;
  targetable: boolean;
  waitTicks: number;
  markedForDeletion: number | false; // -1 -> false, else tick
  health: number | null;
  underConstruction: boolean;
  targetUnitId: number | null;
  targetTile: number | null;
  troops: number;
  missileTimerQueue: number[];
  level: number;
  veterancy: number;
  hasTrainStation: boolean;
  trainType: number | null; // 0=Engine, 1=TailEngine, 2=Carriage
  loaded: boolean | null;
  constructionStartTick: number | null;
}